Move KafkaSchemaRecordType to api package.
