+++ kcbq-api/src/main/java/com/wepay/kafka/connect/bigquery/api/SchemaRetriever.java
@@ -25,7 +25,7 @@ public interface SchemaRetriever {
    * @param schemaType The type of kafka schema, either "value" or "key".
    * @return The Schema for the given table.
    */
-  public Schema retrieveSchema(TableId table, String topic, String schemaType);
+  public Schema retrieveSchema(TableId table, String topic, KafkaSchemaRecordType schemaType);
 
   /**
    * Set the last seen schema for a given topic
