+++ kcbq-confluent/src/main/java/com/wepay/kafka/connect/bigquery/schemaregistry/schemaretriever/SchemaRegistrySchemaRetriever.java
@@ -2,6 +2,7 @@ package com.wepay.kafka.connect.bigquery.schemaregistry.schemaretriever;
 
 import com.google.cloud.bigquery.TableId;
 
+import com.wepay.kafka.connect.bigquery.api.KafkaSchemaRecordType;
 import com.wepay.kafka.connect.bigquery.api.SchemaRetriever;
 
 import io.confluent.connect.avro.AvroData;
@@ -61,7 +62,7 @@ public class SchemaRegistrySchemaRetriever implements SchemaRetriever {
   }
 
   @Override
-  public Schema retrieveSchema(TableId table, String topic, String schemaType) {
+  public Schema retrieveSchema(TableId table, String topic, KafkaSchemaRecordType schemaType) {
     String subject = getSubject(topic, schemaType);
     try {
       logger.debug("Retrieving schema information for topic {} with subject {}", topic, subject);
@@ -80,7 +81,7 @@ public class SchemaRegistrySchemaRetriever implements SchemaRetriever {
   @Override
   public void setLastSeenSchema(TableId table, String topic, Schema schema) { }
 
-  private String getSubject(String topic, String schemaType) {
-    return topic + "-" + schemaType;
+  private String getSubject(String topic, KafkaSchemaRecordType schemaType) {
+    return topic + "-" + schemaType.toString();
   }
 }
